adding proftpd, updated from 10.2 tree;
1.2.9RC2
The changes in this release are as follows:
This release includes numerous bugfixes, including portability fixes
for Cygwin, FreeBSD, and Solaris. The major new feature added in this
release is support for RFC 2428, which defines support for FTP over
IPv6.

1.2.9RC3
The changes in this release are as follows:
Many bugs and a security issue were fixed.

1.2.9
The changes in this release are as follows:
This version now supports IPv6, and adds and improves  some new
modules like "mod_rewrite", "mod_ifsession",  and TLS Support for
encrypted FTP sessions. There are  some new directives for checking
anonymous passwords  in a regular expression, a RootRevoke for
dropping all  root privileges, and lots more.