Adding gproftpd, updated from 10.2, new filename;
8.0.4
The changes in this release are as follows:
A global limit login directive was added for extra security. A top 10
ul/dl statistics generator called gprostats was added that can make
welcome messages for all currently active users and/or generate HTML
statistics via crond or gproftpd. The files tab is now a liststore.

8.0.5
The changes in this release are as follows:
A listing bug in gprostats has been fixed, and a creation date added.
The userchanger has been made more dynamic. A per-user setting has
been added to allow only specified addresses to connect to the
server. Some missing UTF-8 conversions have been added. A popup
question to add Gproftpd's default configuration will now be shown if
proftpd.conf doesnt exist or doesnt have the directives Gproftpd
wants it to have. The way it detects on/offline status has been
changed. FTP-TLS options and several OS defines have been added.

8.0.6
The changes in this release are as follows:
A certificate creation section is now displayed  in the global tab
along with FTP-TLS options if  Proftpd is compiled with mod_tls.
DirFakeUser  and DirFakeGroup were added to the global tab.  The
upload directory and its limits can now be  changed from the user
settings window.

8.0.7
The changes in this release are as follows:
This version cleans up some parts of the code, and adds  path
corrections and more error detections for when  adding users.

8.0.8
The changes in this release are as follows:
The scrolling behaviour in the transfer tab has been fixed. A maximum
allowed length has been added to all input widgets. If some other
server is running on the port being used, and the user presses
offline, it will now inform the user of this. More path corrections
have been added, and it will now show the directories changed to in
the user changer.

8.0.9
The changes in this release are as follows:
It is now possible to update welcome messages on a  per-user basis. A
progress bar is now displayed  during certificate generation. Three
more questions  are asked when installing gproftpd and deleting a
user. A module detection bug was fixed, and users  can now be kicked
and kick-banned by username or  PID. The randomizer was remade and a
banned  column was added to the user list. A SystemLog  directive was
added. The randomize username button  now appends the randomized
username to the end  of the path.

8.1.0
The changes in this release are as follows:
This version fixes a small bug in the kickban function. It  will now
remember the useradd and statistics paths. It  will remember the
randomize length for usernames and  passwords and all can be changed
in the global tab. It will  now add the master ftpusers group if it
doesn't exist. The  timeoutidle directive was added.