New Upstream Release;
8.1.1
The changes in this release are as follows:
This release, codenamed "Christmas stocking", fixes a  uint64
dependency for Darwin, handles split filenames,  includes more
security checking, and features support  for managing multiple
virtual hosts.

8.1.2
The changes in this release are as follows:
This version fixes a problem that arose when the user added an
anonymous directive and there were no users and no vhosts. GProstats
output now shows the default server's name.

8.1.3
The changes in this release are as follows:
This release adds the Bind and SocketBindTight  directives, and adds a
check for /bin/false in shells. The  port can now be set to 0 to turn
a server off temporarily.

8.1.4
The changes in this release are as follows:
Passwords will now be encrypted using the MD5 extension if it exists.
The selection dialogs for the directory paths have been fixed.
Randomized usernames and passwords in upper or lower case can now be
selected.