New upstream release;
The changes in this release are as follows:
The tv_grab_uk grabber has been removed, but tv_grab_uk_rt has been
enhanced a bit as a short-term replacement. There are bugfixes to the
XMLTV.pm library and filter programs. A segfault in tv_imdb on some
platforms and installation on Cygwin have been fixed.