New revision of subversion 0.37.0 (1.0.0-RC1). See <http://svn.collab.net/repos/svn/trunk/CHANGES> for the change log.
The fink package now contains the swig perl (...-swig-pm581) and swig
python (...-swig-py23) bindings. Also the cvs2svn conversion tool can now
be installed directly with the package cvs2svn(-ssl)

As this will be almost the same as 1.0.0 it would be great if people would
test this version thoroughly and report back to me.