New upstream release;
1.7.8
The changes in this release are as follows:
A lasso selection tool has been added. Another bug relating to EPS
output has been fixed. This release requires OpenBabel 1.100.2.

1.7.7
The changes in this release are as follows:
Bugs in EPS output and text alignment have been fixed. The project has
moved to sourceforge.net.