New upstream release;
8.1.6
The changes in this release are as follows:
A user importer was added in the entire tab. The code was restructured
and cleaned up. The MasqueradeAddress and PassivePorts directives
were added. A function was added so that userprofiles and
virtualhosts will be separated with only one newline.

8.1.5
The changes in this release are as follows:
A bug in the xferlog parser's action code has been fixed. Incorrect
salting for users that couldnt link with -lcrypt has been fixed.
configure.in sets the version and reflects proftpd more. Separate
functions for list_disc and list_secure have been added. useradd,
userdel, and passwd support have been added for OS X, which seems to
make the port complete.