New upstream release;
1.41
The changes in this release are as follows:
This release fixes build errors when compiled without --with-db.

1.40
The changes in this release are as follows:
This release adds native support for the Berkeley db format output for
the relay file and support for Postfix. Server configurations have
been broken out to provide much simpler configuration.