New upstream release;
2.12.2
The changes in this release are as follows:
Segfaults in multiple option handling were fixed.  Sections can now be
associated with a more detailed  description that is printed in the
--help output.  Options and arguments in config files can now be
separated by '='. Wrapping is now applied to all  description
strings. Better checks for quoted string  well-formedness in the
configuration files were  implemented.