New upstream release;
1.8.5
The changes in this release are as follows:
EPS support has been improved. Transparent pixmaps are now used to
enhance print and display quality. A number of warnings and
deprecated functions have been fixed.

1.9
The changes in this release are as follows:
Numerous bugs were fixed, including problems with  international
character support and the text and  bond tools.