New upstream release for libcdio and libcddb
libcddb 0.9.6
The changes in this release are as follows:
Libcddb now uses the FreeDB protocol version 6. The difference with
version 5 is that the responses from the FreeDB server are in UTF-8
format instead of ISO8859-1. So if the CD you query contains
non-US-ASCII characters, the results will be different in this
version of libcddb. Support for character set conversion has also
been added. Several bugs have been fixed.