New upstream release;
0.62
The changes in this release are as follows:
Several documentation improvements were made. Fixes were made to the
winbot format. A new lulubot format was added. A new option was added
to make it easier to integrate the results into existing HTML pages.

0.61
The changes in this release are as follows:
This release fixes all known bugs, including a rewritten URL  parser,
a new solution to the problematic mIRC 6 format,  and a fix to a
wrong warning message which occurred when  empty log files were
parsed.