New Upstream Release;
1.9.4
The changes in this release are as follows:
Three-dimensional structure generation and database code have been
updated to give better results. An Insert Symbol dialog is now
available. A few other small bugs have been fixed.

1.9.3
The changes in this release are as follows:
BUILD3D is now accessed via a CGI script, so it should be available to
all users with an Internet connection. Some issues with colors have
been resolved. An issue with save and undo of drawings containing "C"
labels has been resolved. Aliphatic chains are consistently drawn
correctly. A new Polish language translation was contributed.

1.9.2
The changes in this release are as follows:
In this release, bugs in drawing behavior and converting the
coordinate system in external file formats have been fixed.  Some
additional options are available for editing drawings.  Support for
generating 3-D structures using an external  program has been added.

1.9.1
The changes in this release are as follows:
Significant fixes have been made to arrow drawing, text drawing, and
the custom structure menu. More settings are saved between sessions,
and a new German translation is included.