New upstream release;
8.2.2
The changes in this release are as follows:
This version features bugfixes in applying server changes  and listing
server settings.

8.2.1
The changes in this release are as follows:
This release fixed a bug that occurred when TLS was used in version
8.2.0. Individual certificates can now be created for each server.
The docdir version has been removed.

8.2.0
The changes in this release are as follows:
Single-clicks are now used instead of double-clicks in the server
changer. The Consolehelper will now be used if it exists. inttypes.h
is used on Solaris. The install-data-local rule now allows "make
DESTDIR=/foobar install". A bug which occurred when mod_tls wasn't
used was fixed.

8.1.9
The changes in this release are as follows:
This release fixes a security issue in gprostats (the statistics
generator). It adds a Simplified Chinese translation.

8.1.8
The changes in this release are as follows:
This release changes the reuse address to 0, adds the function
allocate(), takes bzero out instead of defining it as memset, and
fixes a zeroing bug in the function xfer_update().