New upstream release, removed patched that were nolonger required.
3.5.4
The changes in this release are as follows:
The setting mirror:include-regex has been added. SSL keys are loaded
for protected data connections as well (as in 3.4 and before). The
new option ftp:ssl-data-use-keys can be used to disable it. The
human-readable options of du and cls work again. Compilation on
certain systems has been fixed.

3.5.2
The changes in this release are as follows:
The program was fixed to strip CRs from ls output, just as before. The
help for pget was updated. A hang in mirror -c and get -c was fixed.
Mirror timestamps for HFTP and HTTP were fixed. Mirror was fixed for
files starting with a tilde. The use of PROPFIND was fixed when
http:use-propfind is to "no". Directory listing and globbing for FTP
over HTTP proxy was fixed. The gnulib sources were updated.

3.5.1
The changes in this release are as follows:
pget -c (continue) with a status file was implemented. New options for
the "repeat" command were added. Listing and DNS caching were
optimized. The ability to enable or disable caching per host was
added. ABOR was made more robust. Support for the ALLO command and
the ftp:use-allo setting was added. Uploading in the fish protocol
was sped up. Two bugs were fixed.