New upstream release, new pkg layout (only xdrawchem depends on it, so I changed it now), upgrades should work fine.
2.0.2
The changes in this release are as follows:
A large number of SMILES and SMARTS bugs were fixed. Several crashes
were fixed, and countless other bugs were fixed.

2.0.1
The changes in this release are as follows:
Highlights include improved memory usage and fixes for compilation
problems on AIX, Solaris, Fedora Core, and with the new GCC-4.1
compiler. Several crashes were fixed, and significantly improved
scripting wrappers for Perl and Python are available. Support for
Cygwin and countless bugfixes are also included.

2.0.0
The changes in this release are as follows:
This release marks a milestone as a stable release for developers and
users alike. Highlights include a new conversion framework that makes
it easier to develop new translators, dramatically improved support
for merging, splitting, and batch conversion, a framework for
molecular fingerprints, similarity searching, a fast molecular
database format, support for Perl and Python scripting "wrappers,"
automatic support for reading .gz (gzip) compressed files, support
for the new IUPAC/NIST InChI identifiers, and more.