New upstream release, enabled sdl support, sound now works
1.5.12
The changes in this release are as follows:
This release corrects a potential exploit against the Windows server,
which was not fully patched by the 1.5.11 release.

1.5.11
The changes in this release are as follows:
This is largely a security bugfix release which fixes a potentially
exploitable bug in the Windows server when running as an NT service.