New upstream release.  New lib versionning.
1.2.1
The changes in this release are as follows:
The full text search feature can now be fine-tuned. Functionality has
been added to specify which fields and categories to search. A bug
due to which the sites command returned random values has also been
fixed.

1.2.0
The changes in this release are as follows:
Full text search, a much requested feature, has been added. This
allows for searching the FreeDB database without actually having any
data about a physical CD.

1.1.0
The changes in this release are as follows:
This release added support for retrieving a list of mirror sites. This
new version is not binary compatible with previous versions, though
the API has not changed and recompiling software using the library
should suffice.