New upstream release;
0.7.23
The changes in this release are as follows:
Some additions to libvcdinfo were made to facilitate handling mouse
"hot spots" in media players. Other small additions, bugfixes, and
documentation updates were also performed.

0.7.22
The changes in this release are as follows:
Some small memory leaks and invalid memory references  were fixed. An
option "notrack" was added to vcdxrip to skip  ripping of a specific
track. A single argument in vcd-info now  acts like the -i option. A
sample program was added to show  the use of libvcdinfo.