New upstream buildbot (0.7.6)
From the release notes:

This release introduces the 'upgrade-master' command, which should be run on
any existing buildmasters you have before restarting them with the new code.
This command will add the public_html directory and its files, and will check
your master.cfg for compatibility problems. This should help to reduce the
downtime and surprises as you upgrade your installations to the new code.