Remove old distributions prior to 10.7
They haven't been supported for years, in some cases over a decade.
If anybody wants to use them, they are free to create a fork of this
repository, with a new branch based before this commit.

Leave 10.7 in, because (a) it is conceivable a few people might still
be using it, and (b) it seems to be sometimes useful to dig through it
when working on the current tree, 10.9-libcxx