Remove Depends:pkgconfig that is no longer needed
We used to need this when .pc used some new syntax to assure a
pkg-config that processed it properly, but no currently supported
distro has ever had a version of pkg-config older than the one that
supports it. Current practice is not to have this dependency solely
because there is a .pc file, and there are other programs now that
could be used instead. Only retain in cases where a program in the
.deb actually runs the pkg-config program.