Update TexLive to 20210324 This bumps kpathsea to 6.3.3, ptexenc to 1.3.9, and teckit to 2.5.10 Make sure updated TexLive (Build)Depends on these new libN
For TexLive, comment out all bits that were used for Japanese formatting.
Unfortunately, most of the accessory code our package added to TexLive was last updated in 2015 or earlier and it's not clear that any of it will still work with the new TexLive.
However, it should be easy(?) to add the code back if it still works on TexLive2021.