Break a dep-cycle: package-stash-pmXXX vs namespace-clean-pmXXX
P::S breaks n::c <= 0.18, so it seems like we had P::S:Depends:n::c to
force beyond it (though the version listed was not quite correct). But
n::c is so far beyond 0.18 now, let's assume nobody still has it that
old or that the breakage itself isn't so bad. More important to not
have actual runtime cycle that prohibits building at all.